Revert "feat: add feedback trace server APIs (#1718)"
This reverts commit f6ae2f40de26adb0cc0e90b6916c27e3598372d4.
