+++ weave/trace_server/errors.py
@@ -8,9 +8,3 @@ class RequestTooLarge(Error):
     """Raised when a request is too large."""
 
     pass
-
-
-class InvalidRequest(Error):
-    """Raised when a request is invalid."""
-
-    pass
