+++ weave/trace_server/sqlite_trace_server.py
@@ -9,9 +9,7 @@ import datetime
 import json
 import hashlib
 import sqlite3
-from zoneinfo import ZoneInfo
 
-import emoji
 
 from .trace_server_interface_util import (
     extract_refs_from_values,
@@ -22,15 +20,7 @@ from . import trace_server_interface as tsi
 from .interface import query as tsi_query
 
 from weave.trace import refs
-from weave.trace_server.emoji_util import detone_emojis
-from weave.trace_server.errors import InvalidRequest
-from weave.trace_server.feedback import (
-    TABLE_FEEDBACK,
-    validate_feedback_create_req,
-    validate_feedback_purge_req,
-)
 from weave.trace_server.trace_server_interface_util import (
-    generate_id,
     WILDCARD_ARTIFACT_VERSION_AND_PATH,
 )
 from weave.trace_server.refs_internal import (
@@ -39,7 +29,6 @@ from weave.trace_server.refs_internal import (
     OBJECT_ATTR_EDGE_NAME,
     TABLE_ROW_ID_EDGE_NAME,
 )
-from weave.trace_server.orm import Row
 
 MAX_FLUSH_COUNT = 10000
 MAX_FLUSH_AGE = 15
@@ -65,14 +54,13 @@ def get_conn_cursor(db_path: str) -> tuple[sqlite3.Connection, sqlite3.Cursor]:
     return conn_cursor
 
 
-class SqliteTraceServer(tsi.TraceServerInterfacePostAuth):
+class SqliteTraceServer(tsi.TraceServerInterface):
     def __init__(self, db_path: str):
         self.lock = threading.Lock()
         self.db_path = db_path
 
     def drop_tables(self) -> None:
         conn, cursor = get_conn_cursor(self.db_path)
-        cursor.execute(TABLE_FEEDBACK.drop_sql())
         cursor.execute("DROP TABLE IF EXISTS calls")
         cursor.execute("DROP TABLE IF EXISTS objects")
         cursor.execute("DROP TABLE IF EXISTS tables")
@@ -144,7 +132,6 @@ class SqliteTraceServer(tsi.TraceServerInterfacePostAuth):
                 val BLOB)
             """
         )
-        cursor.execute(TABLE_FEEDBACK.create_sql())
 
     # Creates a new call
     def call_start(self, req: tsi.CallStartReq) -> tsi.CallStartRes:
@@ -465,12 +452,12 @@ class SqliteTraceServer(tsi.TraceServerInterfacePostAuth):
                 WITH RECURSIVE Descendants AS (
                     SELECT id
                     FROM calls
-                    WHERE project_id = ? AND
-                        deleted_at IS NULL AND
+                    WHERE project_id = ? AND 
+                        deleted_at IS NULL AND 
                         parent_id IN (SELECT id FROM calls WHERE id IN ({}))
-
+                    
                     UNION ALL
-
+                    
                     SELECT c.id
                     FROM calls c
                     JOIN Descendants d ON c.parent_id = d.id
@@ -489,7 +476,7 @@ class SqliteTraceServer(tsi.TraceServerInterfacePostAuth):
             delete_query = """
                 UPDATE calls
                 SET deleted_at = CURRENT_TIMESTAMP
-                WHERE deleted_at is NULL AND
+                WHERE deleted_at is NULL AND 
                     id IN ({})
             """.format(
                 ", ".join("?" * len(all_ids))
@@ -715,79 +702,6 @@ class SqliteTraceServer(tsi.TraceServerInterfacePostAuth):
         if query_result is None:
             raise NotFoundError(f"File {req.digest} not found")
         return tsi.FileContentReadRes(content=query_result[0])
-
-    def feedback_create(
-        self, req: tsi.FeedbackCreateReqForInsert
-    ) -> tsi.FeedbackCreateRes:
-        validate_feedback_create_req(req)
-
-        # Augment emoji with alias.
-        res_payload = {}
-        if req.feedback_type == "wandb.reaction.1":
-            em = req.payload["emoji"]
-            if emoji.emoji_count(em) != 1:
-                raise InvalidRequest(
-                    "Value of emoji key in payload must be exactly one emoji"
-                )
-            req.payload["alias"] = emoji.demojize(em)
-            detoned = detone_emojis(em)
-            req.payload["detoned"] = detoned
-            req.payload["detoned_alias"] = emoji.demojize(detoned)
-            res_payload = req.payload
-
-        feedback_id = generate_id()
-        created_at = datetime.datetime.now(ZoneInfo("UTC"))
-        # TODO: Any validation on weave_ref?
-        payload = json.dumps(req.payload)
-        MAX_PAYLOAD = 1024
-        if len(payload) > MAX_PAYLOAD:
-            raise InvalidRequest("Feedback payload too large")
-        row: Row = {
-            "id": feedback_id,
-            "project_id": req.project_id,
-            "weave_ref": req.weave_ref,
-            "wb_user_id": req.wb_user_id,
-            "creator": req.creator,
-            "feedback_type": req.feedback_type,
-            "payload": payload,
-            "created_at": created_at,
-        }
-        conn, cursor = get_conn_cursor(self.db_path)
-        with self.lock:
-            TABLE_FEEDBACK.insert(row).execute(cursor)
-            conn.commit()
-        return tsi.FeedbackCreateRes(
-            id=feedback_id,
-            created_at=created_at,
-            wb_user_id=req.wb_user_id,
-            payload=res_payload,
-        )
-
-    def feedback_query(self, req: tsi.FeedbackQueryReq) -> tsi.FeedbackQueryRes:
-        conn, cursor = get_conn_cursor(self.db_path)
-        query = TABLE_FEEDBACK.select()
-        query = query.project_id(req.project_id)
-        query = query.fields(req.fields)
-        query = query.where(req.query)
-        query = query.order_by(req.sort_by)
-        query = query.limit(req.limit).offset(req.offset)
-        result = query.execute(cursor)
-        return tsi.FeedbackQueryRes(result=result)
-
-    def feedback_purge(self, req: tsi.FeedbackPurgeReq) -> tsi.FeedbackPurgeRes:
-        # TODO: Instead of passing conditions to DELETE FROM,
-        #       should we select matching ids, and then DELETE FROM WHERE id IN (...)?
-        #       This would allow us to return the number of rows deleted, and complain
-        #       if too many things would be deleted.
-        validate_feedback_purge_req(req)
-        conn, cursor = get_conn_cursor(self.db_path)
-        query = TABLE_FEEDBACK.purge()
-        query = query.project_id(req.project_id)
-        query = query.where(req.query)
-        with self.lock:
-            query.execute(cursor)
-            conn.commit()
-        return tsi.FeedbackPurgeRes()
 
     def _table_query(
         self,
