+++ weave/trace_server/trace_server_interface.py
@@ -1,10 +1,9 @@
 import abc
 import datetime
 import typing
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
-
-from .interface.query import Query, LiteralOperation, GetFieldOperator
+from .interface.query import Query
 
 
 class CallSchema(BaseModel):
@@ -319,74 +318,6 @@ class FileContentReadRes(BaseModel):
     content: bytes
 
 
-class FeedbackPayloadReactionReq(BaseModel):
-    emoji: str
-
-
-class FeedbackPayloadNoteReq(BaseModel):
-    note: str = Field(min_length=1, max_length=1024)
-
-
-class FeedbackCreateReq(BaseModel):
-    project_id: str = Field(examples=["entity/project"])
-    weave_ref: str = Field(examples=["weave:///entity/project/object/name:digest"])
-    creator: typing.Optional[str] = Field(default=None, examples=["Jane Smith"])
-    feedback_type: str = Field(examples=["custom"])
-    payload: typing.Dict[str, typing.Any] = Field(
-        examples=[
-            {
-                "key": "value",
-            }
-        ]
-    )
-
-
-class FeedbackCreateReqForInsert(FeedbackCreateReq):
-    wb_user_id: str
-
-
-# The response provides the additional fields needed to convert a request
-# into a complete Feedback.
-class FeedbackCreateRes(BaseModel):
-    id: str
-    created_at: datetime.datetime
-    wb_user_id: str
-    payload: typing.Dict[str, typing.Any]  # If not empty, replace payload
-
-
-class Feedback(FeedbackCreateReqForInsert):
-    id: str
-    created_at: datetime.datetime
-
-
-class FeedbackQueryReq(BaseModel):
-    project_id: str = Field(examples=["entity/project"])
-    fields: typing.Optional[list[str]] = Field(
-        default=None, examples=[["id", "feedback_type", "payload.note"]]
-    )
-    query: typing.Optional[Query] = None
-    # TODO: I think I would prefer to call this order_by to match SQL, but this is what calls API uses
-    # TODO: Might be nice to have shortcut for single field and implied ASC direction
-    # TODO: I think _SortBy shouldn't have leading underscore
-    sort_by: typing.Optional[typing.List[_SortBy]] = None
-    limit: typing.Optional[int] = Field(default=None, examples=[10])
-    offset: typing.Optional[int] = Field(default=None, examples=[0])
-
-
-class FeedbackQueryRes(BaseModel):
-    # Note: this is not a list of Feedback because user can request any fields.
-    result: list[dict[str, typing.Any]]
-
-
-class FeedbackPurgeReq(BaseModel):
-    project_id: str = Field(examples=["entity/project"])
-    query: Query
-
-
-class FeedbackPurgeRes(BaseModel):
-    pass
-
-
 class TraceServerInterface:
     def ensure_project_exists(self, entity: str, project: str) -> None:
         pass
@@ -460,22 +391,4 @@ class TraceServerInterface:
 
     @abc.abstractmethod
     def file_content_read(self, req: FileContentReadReq) -> FileContentReadRes:
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def feedback_create(self, req: FeedbackCreateReq) -> FeedbackCreateRes:
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def feedback_query(self, req: FeedbackQueryReq) -> FeedbackQueryRes:
-        raise NotImplementedError()
-
-    @abc.abstractmethod
-    def feedback_purge(self, req: FeedbackPurgeReq) -> FeedbackPurgeRes:
-        raise NotImplementedError()
-
-
-class TraceServerInterfacePostAuth(TraceServerInterface):
-    @abc.abstractmethod
-    def feedback_create(self, req: FeedbackCreateReqForInsert) -> FeedbackCreateRes:
         raise NotImplementedError()
